% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfSchemes.R
\name{addSchemes}
\alias{addSchemes}
\title{add user defined scheme to an existing \code{\link{SetOfSchemes-class}}
 or extend an existing set of schemes object with additional schemes}
\usage{
addSchemes(setOfSchemes, extraSchemes)
}
\arguments{
\item{setOfSchemes}{\code{\link{SetOfSchemes-class}} object or a matrix of indiviudual schemes}

\item{extraSchemes}{array of schemes to add, see code{\link{SetOfSchemes-class}}}
}
\description{
add user defined scheme to an existing \code{\link{SetOfSchemes-class}}
 or extend an existing set of schemes object with additional schemes
}
