% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraintFunctions.R
\name{check_scheme_exactNumberObsPerTimePoint}
\alias{check_scheme_exactNumberObsPerTimePoint}
\title{check the number of observations per time points is equal specified value}
\usage{
check_scheme_exactNumberObsPerTimePoint(scheme, value)
}
\arguments{
\item{scheme}{a microsampling scheme}

\item{value}{numeric constant}
}
\description{
check the number of observations per time points is equal specified value
}
