% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraintFunctions.R
\name{check_scheme_minObsPerTimePoint}
\alias{check_scheme_minObsPerTimePoint}
\title{check the mimimum observations per time points is above a specified value}
\usage{
check_scheme_minObsPerTimePoint(scheme, value)
}
\arguments{
\item{scheme}{a microsampling scheme}

\item{value}{numeric constant}
}
\description{
check the mimimum observations per time points is above a specified value
}
