% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraintFunctions.R
\name{check_subject_maxConsecSamples}
\alias{check_subject_maxConsecSamples}
\title{check the maximum of consecutive samples per subject falls below the specified value}
\usage{
check_subject_maxConsecSamples(subjectScheme, value)
}
\arguments{
\item{subjectScheme}{a one subject scheme, one line of a scheme}

\item{value}{to compare scheme with}
}
\description{
check the maximum of consecutive samples per subject falls below the specified value
}
