% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfSchemes.R
\name{constructSetOfSchemes}
\alias{constructSetOfSchemes}
\title{construct user defined \code{\link{SetOfSchemes-class}}}
\usage{
constructSetOfSchemes(schemes, timePoints)
}
\arguments{
\item{schemes}{array representing \code{.Data} slot of \code{\link{SetOfSchemes-class}}}

\item{timePoints}{numeric vector, \code{timePoinst} slot of \code{\link{SetOfSchemes-class}}}
}
\description{
construct user defined \code{\link{SetOfSchemes-class}}
}
\examples{
  schemes                      <-  getData( getExampleSetOfSchemes() ) 
  timePoints                   <-  exp(1:4) 
  constructSetOfSchemes( schemes , timePoints)
}
