% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankScheme.R
\name{estimatePopCurve}
\alias{estimatePopCurve}
\title{estimate population Pk curve from a scheme and pkData}
\usage{
estimatePopCurve(scheme, dataArray)
}
\arguments{
\item{scheme}{one scheme from \code{\link{SetOfSchemes-class}}}

\item{dataArray}{\code{\link{PkData-class}}}
}
\description{
estimate population Pk curve from a scheme and pkData
}
\keyword{internal}
