% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemeGenerator.R
\name{flagSchemesMeetingConstraints}
\alias{flagSchemesMeetingConstraints}
\title{internal function to check scheme constraints ( faster )}
\usage{
flagSchemesMeetingConstraints(dataSchemes, schemeConstraints)
}
\arguments{
\item{dataSchemes}{data array for schemes}

\item{schemeConstraints}{data frame with schemeConstraints}
}
\description{
internal function to check scheme constraints ( faster )
}
\keyword{internal}
