% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{flattenSetOfSchemes}
\alias{flattenSetOfSchemes}
\title{Transform 3 way array to 2 way array}
\usage{
flattenSetOfSchemes(object)
}
\arguments{
\item{object}{\code{\link{SetOfSchemes-class}}}
}
\description{
Transform 3 way array to 2 way array
}
