% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{formatTimePoints}
\alias{formatTimePoints}
\title{Format time points as a set}
\usage{
formatTimePoints(timePoints)
}
\arguments{
\item{timePoints}{numeric vector of timme points}
}
\description{
Format time points as a set
}
