% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{genMVN}
\alias{genMVN}
\title{Internal function to generate multivariate data via a cholesky decompostion ( to have replicability accross systems )}
\usage{
genMVN(n, mu, Sigma)
}
\arguments{
\item{n}{number of sample to take}

\item{mu}{mean vector}

\item{Sigma}{covariance matrix}
}
\description{
Internal function to generate multivariate data via a cholesky decompostion ( to have replicability accross systems )
}
\keyword{internal}
