% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{getCoeffVariationError}
\alias{getCoeffVariationError}
\alias{getCoeffVariationError,PkModelParent-method}
\title{generic function to extract \code{coeffVariationError} slot}
\usage{
getCoeffVariationError(object, ...)

\S4method{getCoeffVariationError}{PkModelParent}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{coeffVariationError} slot
}
