% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraintFunctions.R
\name{getConstraintsExample}
\alias{getConstraintsExample}
\title{get a minimal example of a constraint data frame}
\usage{
getConstraintsExample()
}
\description{
get a minimal example of a constraint data frame
}
