% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\alias{getCorrelationMatrix,PkModelParent-method}
\title{generic function to extract the \code{correlationMatrix}-slot}
\usage{
getCorrelationMatrix(object, ...)

\S4method{getCorrelationMatrix}{PkModelParent}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract the \code{correlationMatrix}-slot
}
