% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R,
%   R/objectPkModel.R, R/objectSetOfTimePoints.R
\docType{methods}
\name{getData}
\alias{getData}
\alias{getData,SetOfSchemes-method}
\alias{getData,PkData-method}
\alias{getData,SetOfTimePoints-method}
\title{generic function to extract the \code{.Data}-slot}
\usage{
getData(object, ...)

\S4method{getData}{SetOfSchemes}(object)

\S4method{getData}{PkData}(object)

\S4method{getData}{SetOfTimePoints}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract the \code{.Data}-slot
}
