% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelRange.R
\name{getDevRankingSettings}
\alias{getDevRankingSettings}
\title{function default ranking settings}
\usage{
getDevRankingSettings(objectClass)
}
\description{
@param class you want to get the default ranking settings for
}
\keyword{internal}
