% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{getDosingInfo}
\alias{getDosingInfo}
\alias{getDosingInfo,PkModelParent-method}
\title{generic function to extract \code{dosingInfo}-slot}
\usage{
getDosingInfo(object, ...)

\S4method{getDosingInfo}{PkModelParent}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{dosingInfo}-slot
}
