% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModel.R
\name{getExampleData}
\alias{getExampleData}
\title{generate an mimimal example of a Pk data without a model}
\usage{
getExampleData()
}
\description{
generate an mimimal example of a Pk data without a model
}
\examples{
getExampleData()
}
