% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankScheme.R
\name{getExampleObjective}
\alias{getExampleObjective}
\title{example objective function for \code{\link{rankObject}}}
\usage{
getExampleObjective()
}
\description{
example objective function for \code{\link{rankObject}}
}
