% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkFunctions.R
\name{getExampleParameters}
\alias{getExampleParameters}
\title{get example parameters to use in \code{\link{pkOdeModel2Compartments}} example}
\usage{
getExampleParameters()
}
\description{
get example parameters to use in \code{\link{pkOdeModel2Compartments}} example
}
