% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemeStatistics.R
\name{getExamplePkCurve}
\alias{getExamplePkCurve}
\title{example of 1 pk curve to be used to test pkCurveStat_[function]}
\usage{
getExamplePkCurve(times)
}
\arguments{
\item{times}{a numeric vector of timePoints}
}
\value{
a \code{data.frame} with time and concentration as columns
}
\description{
example of 1 pk curve to be used to test pkCurveStat_[function]
}
\examples{
getExamplePkCurve( times = 0:10  )
}
