% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModel.R
\name{getExamplePkModel}
\alias{getExamplePkModel}
\title{get minimal example of \code{\link{PkModel-class}}}
\usage{
getExamplePkModel()
}
\description{
get minimal example of \code{\link{PkModel-class}}
}
\examples{
getExamplePkModel()
}
