% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelRange.R
\name{getExamplePkModelRange}
\alias{getExamplePkModelRange}
\title{get minimal example of \code{\link{PkModelRange-class}}}
\usage{
getExamplePkModelRange()
}
\description{
get minimal example of \code{\link{PkModelRange-class}}
}
\examples{
getExamplePkModelRange()
}
