% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfSchemes.R
\name{getExampleSetOfSchemes}
\alias{getExampleSetOfSchemes}
\title{get a minimal example of a set of schemes object}
\usage{
getExampleSetOfSchemes()
}
\description{
get a minimal example of a set of schemes object
}
