% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfTimePoints.R
\name{getExampleSetOfTimePoints}
\alias{getExampleSetOfTimePoints}
\title{get a minimal example set of time points to test functions with}
\usage{
getExampleSetOfTimePoints(fullTimePoints, nTimePointsSelect = 5,
  nChoicesSubset = 7)
}
\arguments{
\item{fullTimePoints}{numeric vector of time points}

\item{nTimePointsSelect}{number of time points to select from the full time points, defaults to 5}

\item{nChoicesSubset}{number of all selection to retain for the example to avoid a large object defaults to 7}
}
\description{
get a minimal example set of time points to test functions with
}
\examples{
getExampleSetOfTimePoints( fullTimePoints = 0:10, nTimePointsSelect = 5, nChoicesSubset = 7 ) 
}
