% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankTimePoints.R
\name{getExampleTimeData}
\alias{getExampleTimeData}
\title{generate example PkData object to be used in example rankTimePoints}
\usage{
getExampleTimeData()
}
\description{
generate example PkData object to be used in example rankTimePoints
}
