% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePointGeneration.R
\name{getExampleTimeZones}
\alias{getExampleTimeZones}
\title{working example time zone dataframe to use in examples}
\usage{
getExampleTimeZones()
}
\description{
working example time zone dataframe to use in examples
}
