% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{getModelFunction}
\alias{getModelFunction}
\alias{getModelFunction,PkModelParent-method}
\title{generic function to extract \code{modelFunction} slot from S4-class object}
\usage{
getModelFunction(object, ...)

\S4method{getModelFunction}{PkModelParent}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{modelFunction} slot from S4-class object
}
