% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R,
%   R/objectSetOfTimePoints.R
\docType{methods}
\name{getNames}
\alias{getNames}
\alias{getNames,SetOfSchemes-method}
\alias{getNames,SetOfTimePoints-method}
\title{generic function extract the names of an S4-object}
\usage{
getNames(object, ...)

\S4method{getNames}{SetOfSchemes}(object)

\S4method{getNames}{SetOfTimePoints}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function extract the names of an S4-object
}
