% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R
\docType{methods}
\name{getNSchemes}
\alias{getNSchemes}
\alias{getNSchemes,SetOfSchemes-method}
\title{generic function to extract \code{nSchemes}-slot}
\usage{
getNSchemes(object, ...)

\S4method{getNSchemes}{SetOfSchemes}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{nSchemes}-slot
}
