% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R
\docType{methods}
\name{getNSubjects}
\alias{getNSubjects}
\alias{getNSubjects,SetOfSchemes-method}
\title{generic function to extract \code{nSubjects}-slot}
\usage{
getNSubjects(object, ...)

\S4method{getNSubjects}{SetOfSchemes}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{nSubjects}-slot
}
