% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{getParameters}
\alias{getParameters}
\alias{getParameters,PkModelParent-method}
\title{generic function to extract \code{parameter}-slot}
\usage{
getParameters(object, ...)

\S4method{getParameters}{PkModelParent}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{parameter}-slot
}
