% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModel.R
\docType{methods}
\name{getPkModel}
\alias{getPkModel}
\alias{getPkModel,PkData-method}
\title{generic function extract a \code{\link{PkModel-class}}}
\usage{
getPkModel(object, ...)

\S4method{getPkModel}{PkData}(object)
}
\arguments{
\item{object}{an S4 object}

\item{...}{additional parameters}
}
\description{
generic function extract a \code{\link{PkModel-class}}
}
\examples{
getPkModel( getExampleData() )
}
