% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkFunctions.R
\name{getPkModelArticle}
\alias{getPkModelArticle}
\title{reproduce the example of the article of Helen Barnet et al.}
\usage{
getPkModelArticle()
}
\description{
reproduce the example of the article of Helen Barnet et al.
}
\note{
this models serves only to reproduce results of the article, 
and allows only 1 dose administered at time 0.
}
\examples{
  model       <-  getPkModelArticle()
  summary( model )
  testData    <-    getPkData( model , 1:12 , nSubjectsPerScheme = 3 , nSamples = 7 )
  plotObject( model , times = 0:12  )
   plotAverageRat( model , doseZero = 100 , timePoints = seq(0,12,0.5) )
}
