% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelRange.R
\name{getPkModels}
\alias{getPkModels}
\title{Generate all possible \code{\link{PkModel-class}} from \code{\link{PkModelRange-class}} combination of ranges}
\usage{
getPkModels(object, outputDirectory = NULL)
}
\arguments{
\item{object}{\code{\link{PkModelRange-class}}}

\item{outputDirectory}{directory to save models as \code{.Rds}  objects, defaults to \code{NULL}
when a temporary directory is made to save models}
}
\value{
\code{\link{PkModelRange-class}} objects saved as a subsdirectory of the \code{outputdirectory}
}
\description{
Generate all possible \code{\link{PkModel-class}} from \code{\link{PkModelRange-class}} combination of ranges
}
\note{
the \code{outputDirectory} is should be empty
}
