% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R,
%   R/objectSetOfTimePoints.R
\docType{methods}
\name{getRanking}
\alias{getRanking}
\alias{getRanking,SetOfSchemes-method}
\alias{getRanking,SetOfTimePoints-method}
\title{generic function to extract the  \code{ranking}-slot}
\usage{
getRanking(object, ...)

\S4method{getRanking}{SetOfSchemes}(object)

\S4method{getRanking}{SetOfTimePoints}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract the  \code{ranking}-slot
}
