% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankScheme.R
\name{getResultPerScheme}
\alias{getResultPerScheme}
\title{get a result of the criteria  per scheme zero concentration at time zero}
\usage{
getResultPerScheme(scheme, dataForSchemesInput, getResultFor1Curve,
  varianceMeasure)
}
\arguments{
\item{scheme}{one scheme of \code{\link{SetOfSchemes-class}}}

\item{dataForSchemesInput}{\code{\link{PkData-class}}}

\item{varianceMeasure}{a function such as \code{var}}

\item{criteria}{a character vector of criteria such as \code{auc} see \code{\link{pkCurveStat}}}
}
\description{
get a result of the criteria  per scheme zero concentration at time zero
}
\keyword{internal}
