% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelRange.R
\name{getSummaryRanks}
\alias{getSummaryRanks}
\title{internal function to rank by a summary function over criterio}
\usage{
getSummaryRanks(rankingOverview, names, summaryFunction)
}
\arguments{
\item{rankingOverview}{matrix with columns scenarios and columns objects names to rank}

\item{names}{of the objects to return}

\item{summaryFunction}{function to apply over scenarios}
}
\value{
data.frame with name and criterion ordered
}
\description{
internal function to rank by a summary function over criterio
}
\keyword{internal}
