% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R,
%   R/objectPkModel.R, R/objectSetOfTimePoints.R
\docType{methods}
\name{getTimePoints}
\alias{getTimePoints}
\alias{getTimePoints,SetOfSchemes-method}
\alias{getTimePoints,PkData-method}
\alias{getTimePoints,SetOfTimePoints-method}
\title{generic function to extract \code{timePoints}-slot}
\usage{
getTimePoints(object, ...)

\S4method{getTimePoints}{SetOfSchemes}(object)

\S4method{getTimePoints}{PkData}(object)

\S4method{getTimePoints}{SetOfTimePoints}(object)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}
}
\description{
generic function to extract \code{timePoints}-slot
}
