% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalHelpers.R
\name{getTopNRanking}
\alias{getTopNRanking}
\title{extract the top n rankings as numeric vector}
\usage{
getTopNRanking(ranking, nSelect, top = TRUE)
}
\arguments{
\item{ranking}{ranking slot of a \code{\link{SetOfTimePoints-class}} or \code{\link{SetOfSchemes-class}}}

\item{nSelect}{the number of items to select}

\item{top}{logical value if \code{TRUE} the top of the ranking is selected, if \code{FALSE} the bottom of the ranking is selected,
defaults to  \code{TRUE}}
}
\value{
numeric vector of items (number of timePointOption or scheme ) from highest to lowest rank
}
\description{
extract the top n rankings as numeric vector
}
