% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePointGeneration.R
\name{\%ARC\%}
\alias{\%ARC\%}
\title{All Row Combinations (ARC) function 
take all combination of rows of 2 matrices and bind them together}
\usage{
matrix1 \%ARC\% matrix2
}
\arguments{
\item{matrix1}{numeric matrix}

\item{matrix2}{numeric matrix}
}
\value{
numeric matrix
}
\description{
All Row Combinations (ARC) function 
take all combination of rows of 2 matrices and bind them together
}
