% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkFunctions.R
\name{oneCompartmentOralModel}
\alias{oneCompartmentOralModel}
\title{solution of one compartmental oral administration model
only use one set of parameters, times can input can be an numeric array}
\usage{
oneCompartmentOralModel(parameters, time, dosingInfo)
}
\arguments{
\item{parameters}{a numeric verctor of parameters as input to the model with names
  \itemize{
    \item Ka: constant absorption rate
    \item Ke: constant elimination rate
    \item dose: initial dose
    \item volume: volume to which the dose is administered
}}

\item{time}{a numeric vector containing timePoints at which the concentration should be predicted
timepoint zero is defined as the moment the dose is administered}

\item{dosingInfo}{see \code{link{PkModel-class}} but opnly one dose at time zero allowed}
}
\value{
vector of concentrations corresponding to the input timePoints
}
\description{
solution of one compartmental oral administration model
only use one set of parameters, times can input can be an numeric array
}
