% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelParent.R
\docType{class}
\name{PkModelParent-class}
\alias{PkModelParent-class}
\title{virtual class, for testing \code{\link{PkModel-class}} and \code{\link{PkModelRange-class}} together}
\description{
virtual class, for testing \code{\link{PkModel-class}} and \code{\link{PkModelRange-class}} together
}
\section{Slots}{

\describe{
\item{\code{modelFunction}}{a function of parameters and hyperparameters}

\item{\code{parameters}}{a data.frame of parameters of mean parameters as input to the modelFunction
with columns:
    \itemize{
      \item parameter: parameter name for 
      \item explanation: optional explantion
      \item value: fixed parameter value for \code{\link{PkModel-class}},
      for \code{\link{PkModel-class}} split up between \code{minValue} and \code{maxValue}               
      \item coeffVariation: the coefficient of variation (  standard deviation / mean ) specifying between-subject variation,
      for \code{\link{PkModelRange-class}}  split up into \code{minValue} and \code{maxValue}  
    }}

\item{\code{correlationMatrix}}{correlation matrix of parameters at the log-scale}

\item{\code{coeffVariationError}}{the coefficient of variation for residual normally distributed error, 
  for \code{\link{PkModelRange-class}}  split up into \code{minCoeffVariationError} and \code{maxCoeffVariationError} 
@slot dosingInfo data frame containing:
      \itemize{
        \item time numeric,  times when a dose is administered
        \item dose numeric, with an amount of dose 
      }}
}}

\author{
Adriaan Blommaert
}
\keyword{internal}
