% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkFunctions.R
\name{pkOdeModel2Compartments}
\alias{pkOdeModel2Compartments}
\title{Set of differential equations representing two compartmental pk model with possible Michaelis-Menten kinetics at absorption and clearance
used as interal function}
\usage{
pkOdeModel2Compartments(t, y, parameters)
}
\arguments{
\item{t}{numeric representing time in the system}

\item{y}{list of state variables ( dose in gut , concentration in plasma , concentration in tissue )}
}
\description{
Set of differential equations representing two compartmental pk model with possible Michaelis-Menten kinetics at absorption and clearance
used as interal function
}
\keyword{internal}
