% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkFunctions.R
\name{plotAverageRat}
\alias{plotAverageRat}
\title{plot plasma concentration for average individual (i.e average parameter values) in function of dose at time zero}
\usage{
plotAverageRat(pkModel, doseZero, timePoints)
}
\arguments{
\item{pkModel}{\code{\link{PkModel-class}}}

\item{doseZero}{numeric value, dose given at time zero}

\item{timePoints}{a numeric vector of time points to plot the plasma concentration at}
}
\value{
ggplot object
}
\description{
plot plasma concentration for average individual (i.e average parameter values) in function of dose at time zero
}
\note{
dose inside de \code{pkModel} is not used
}
\examples{
 plotAverageRat( getExamplePkModel() , 2 , seq( 0 , 20, 0.1 ) )
}
