% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelRange.R
\name{rankBasedOnDirectory}
\alias{rankBasedOnDirectory}
\title{internal function apply ranking per directory}
\usage{
rankBasedOnDirectory(objectToRank, subDir, nSimulations, rankingSettings,
  nCores = 1)
}
\description{
internal function apply ranking per directory
}
\note{
use version of rankObject without defaults
}
\keyword{internal}
