% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{runMicrosamplingDesignApp}
\alias{runMicrosamplingDesignApp}
\title{Run the PkPlotApp in a new tab in your browser}
\usage{
runMicrosamplingDesignApp(installDependencies = FALSE, ...)
}
\arguments{
\item{installDependencies}{boolean, whether to first install packages listed
in the Suggests field of DESCRIPTION; default value is FALSE}

\item{...}{further arguments that can be passed to \code{\link[shiny]{runApp}}}
}
\value{
no return value

no return value
}
\description{
Run the PkPlotApp in a new tab in your browser
}
\note{
on windows computer, nCores set to 1 because forking not possible on windows machine

The shiny application contains a 2 compartmental oral dosing pharmacokinetic model which does not include measurement error
}
\keyword{Run}
\keyword{application}
\keyword{bestShiny}
\keyword{internal}
\keyword{the}
