% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{setCoeffVariationError<-}
\alias{setCoeffVariationError<-}
\alias{setCoeffVariationError<-,PkModelParent-method}
\title{replace \code{coeffVariationError}-slot}
\usage{
setCoeffVariationError( object ) <-  value

\S4method{setCoeffVariationError}{PkModelParent}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a value containing the coefficient of variation of the error term}
}
\description{
replace \code{coeffVariationError}-slot
}
