% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{setCorrelationMatrix<-}
\alias{setCorrelationMatrix<-}
\alias{setCorrelationMatrix<-,PkModelParent-method}
\title{replace \code{correlationMatrix}-slot}
\usage{
setCorrelationMatrix( object ) <-  value

\S4method{setCorrelationMatrix}{PkModelParent}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a matrix containing correlations between parameters}
}
\description{
replace \code{correlationMatrix}-slot
}
