% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{setDosingInfo<-}
\alias{setDosingInfo<-}
\alias{setDosingInfo<-,PkModelParent-method}
\title{replace \code{dosingInfo}-slot}
\usage{
setDosingInfo(object) <-  value

\S4method{setDosingInfo}{PkModelParent}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a data.frame containing dosing information}
}
\description{
replace \code{dosingInfo}-slot
}
