% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkFunctions.R
\name{setModelToAverageRat}
\alias{setModelToAverageRat}
\title{get a model with all variances to zero}
\usage{
setModelToAverageRat(pkModel)
}
\arguments{
\item{pkModel}{\code{\link{PkModel-class}}}
}
\description{
get a model with all variances to zero
}
