% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfSchemes.R
\docType{class}
\name{SetOfSchemes-class}
\alias{SetOfSchemes-class}
\alias{setOfSchemes}
\alias{SetOfSchemes}
\title{S4 class SetOfSchemes representing a set of designs with given time points}
\description{
S4 class SetOfSchemes representing a set of designs with given time points
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a logical array of 3 dimensions ( nSubjects x nTimePoints x nSchemes )}

\item{\code{timePoints}}{numeric vector of time Points}

\item{\code{nSchemes}}{integer value number of schemes}

\item{\code{nSubjects}}{numeric maximum number of subjects per scheme}

\item{\code{designConstraints}}{a data.frame of constraints on possible sampling schemes as background information}

\item{\code{ranking}}{is a data.frame which is the rank of the schemes according to a specific criterion}
}}

\author{
Adriaan Blommaert
}
