% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfTimePoints.R
\docType{class}
\name{SetOfTimePoints-class}
\alias{SetOfTimePoints-class}
\alias{SetOfTimePoints}
\alias{setOfTimePoints}
\title{S4 class SetOfTimePoints representing a set of designs with given time points}
\description{
S4 class SetOfTimePoints representing a set of designs with given time points
}
\section{Slots}{

\describe{
\item{\code{.Data}}{a numerics array of 2 dimensions ( nTimePointChoices  x nTimePointsSelect) contains per time point choice the selected time points in hours}

\item{\code{fullTimePoints}}{numeric vector of all time points one is willing to consider}

\item{\code{nFullTimePoints}}{number of all time points one is willing to consider}

\item{\code{nTimePointsSelect}}{number of time points selected from the fullTimePoints}

\item{\code{nTimePointOptions}}{number of possible timePoint choices}

\item{\code{ranking}}{is a data.frame which is the rank of the timePointChoices according to a specific criterion.}
}}

\author{
Adriaan Blommaert
}
