% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModelParent.R
\docType{methods}
\name{setParameters<-}
\alias{setParameters<-}
\alias{setParameters<-,PkModelParent-method}
\title{replace \code{parameters}-slot}
\usage{
setParameters( object ) <-  value

\S4method{setParameters}{PkModelParent}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a data.frame containing parameters}
}
\description{
replace \code{parameters}-slot
}
