% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R,
%   R/objectSetOfTimePoints.R
\docType{methods}
\name{setRanking<-}
\alias{setRanking<-}
\alias{setRanking<-,SetOfSchemes-method}
\alias{setRanking<-,SetOfTimePoints-method}
\title{replace \code{ranking}-slot}
\usage{
setRanking(object) <-  value

\S4method{setRanking}{SetOfSchemes}(object) <- value

\S4method{setRanking}{SetOfTimePoints}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a data.frame containing a ranking}
}
\description{
replace \code{ranking}-slot
}
