% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectSetOfSchemes.R
\docType{methods}
\name{setTimePoints<-}
\alias{setTimePoints<-}
\alias{setTimePoints<-,SetOfSchemes-method}
\title{generic function to replace \code{timePoints}-slot}
\usage{
setTimePoints(object) <- value

\S4method{setTimePoints}{SetOfSchemes}(object) <- value
}
\arguments{
\item{object}{a S4 class object}

\item{value}{a vector of time points}
}
\description{
generic function to replace \code{timePoints}-slot
}
