% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaaGenerics.R, R/objectPkModel.R
\docType{methods}
\name{subsetOnTimePoints}
\alias{subsetOnTimePoints}
\alias{subsetOnTimePoints,PkModel-method}
\title{generic function to subset the \code{timePoints}-slot and generate an object of the same class}
\usage{
subsetOnTimePoints(object, ...)

\S4method{subsetOnTimePoints}{PkModel}(object, timePointsSelect)
}
\arguments{
\item{object}{a S4 class object}

\item{...}{additional parameters}

\item{timePointsSelect}{a subset of time points to select data for}
}
\description{
generic function to subset the \code{timePoints}-slot and generate an object of the same class
}
\examples{
\dontshow{ pkData  <- getExampleData() }
subsetOnTimePoints( pkData, c( 1 , 2 ) )
subsetOnTimePoints( object = pkData, timePointsSelect =  c( 1 , 2 ) )
}
