% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectSetOfSchemes.R
\docType{methods}
\name{summary,SetOfSchemes-method}
\alias{summary,SetOfSchemes-method}
\title{summarize object}
\usage{
\S4method{summary}{SetOfSchemes}(object, printToConsole = TRUE)
}
\arguments{
\item{object}{\code{\link{SetOfSchemes-class}}}

\item{printToConsole}{logical value if \code{TRUE} prints to console ,
if \code{FALSE} outputs text element ,
defaults to \code{TRUE}}
}
\description{
summarize object
}
