% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectPkModelParent.R
\docType{methods}
\name{summary,PkModelParent-method}
\alias{summary,PkModelParent-method}
\title{function to summarize an object}
\usage{
\S4method{summary}{PkModelParent}(object)
}
\arguments{
\item{object}{\code{\link{PkModel-class}}}
}
\description{
function to summarize an object
}
